requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

flask-shell-ipython==0.2.2

# Testing
pytest==3.0.5
WebTest==2.0.24
factory-boy==2.6.1
flex==6.3.0

# Lint and code style
flake8==3.2.1
flake8-blind-except==0.1.1
flake8-debugger==1.4.0
flake8-docstrings==1.0.2
flake8-isort==2.1.3
flake8-quotes==0.8.1
isort==4.2.2
pep8-naming==0.4.1

bandit==1.3.0
codeclimate-test-reporter==0.2.0
pytest-cov==2.4.0

jsonpickle==0.9.5
